+++ packages/pn-commons/src/components/NotificationDetail/CostDetailsDrawerContent/UnavailableData.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
+import { Stack, Typography } from '@mui/material';
+
+import { getLocalizedOrDefaultLabel } from '../../../utility/localization.utility';
+
+const UnavailableDataDrawerContent: React.FC = () => (
+  <Stack direction="row" alignItems="center" spacing={2}>
+    <HelpOutlineIcon fontSize="small" sx={{ flexShrink: 0, color: '#BBC2D6' }} />
+    <Typography variant="body2" fontSize="14px" color="#555C70">
+      {getLocalizedOrDefaultLabel('notifications', 'notification-alert.details.more-info')}
+    </Typography>
+  </Stack>
+);
+
+export default UnavailableDataDrawerContent;