+++ packages/pn-commons/src/utility/notification.utility.ts
@@ -11,7 +11,6 @@ import {
   F24PaymentDetails,
   INotificationDetailTimeline,
   LegalFactType,
-  NotificationDeliveryMode,
   NotificationDetailDocument,
   NotificationDetailOtherDocument,
   NotificationDetailPayment,
@@ -32,6 +31,16 @@ import { getLocalizedOrDefaultLabel } from '../utility/localization.utility';
 import { TimelineStepInfo } from './TimelineUtils/TimelineStep';
 import { TimelineStepFactory } from './TimelineUtils/TimelineStepFactory';
 
+type StatusInfo = {
+  label: string;
+  tooltip: string;
+  description: string;
+};
+
+const AnalogDeliveryCodeStock = new Set(['RECRN003C', 'RECRN011', 'RECAG011A']);
+const AnalogDeliveryCodeWithdrawnStock = new Set(['RECAG005C', 'RECAG006C']);
+const AnalogDeliveryCodeExpiredStock = new Set(['RECAG008C', 'RECRN005C', 'PNAG012', 'PNRN012']);
+
 /*
  * Besides the values used in the generation of the final messages,
  * data can include an isMultiRecipient attribute, which refers to the notification.
@@ -39,17 +48,7 @@ import { TimelineStepFactory } from './TimelineUtils/TimelineStepFactory';
  * (instead of just "-tooltip" and "-description")
  * entries will be looked for in the i18n catalog.
  */
-function localizeStatus(
-  status: string,
-  defaultLabel: string,
-  defaultTooltip: string,
-  defaultDescription: string,
-  data?: { [key: string]: any }
-): {
-  label: string;
-  tooltip: string;
-  description: string;
-} {
+function localizeStatus(status: string, data?: { [key: string]: any }): StatusInfo {
   const isMultiRecipient = data?.isMultiRecipient;
   // eslint-disable-next-line functional/no-let
   let filteredData: any = omit(data, ['isMultiRecipient']);
@@ -60,22 +59,116 @@ function localizeStatus(
   return {
     label: getLocalizedOrDefaultLabel(
       'notifications',
-      `status.${status}${isMultiRecipient ? '-multirecipient' : ''}`,
-      defaultLabel
+      `status.${status}${isMultiRecipient ? '-multirecipient' : ''}`
     ),
     tooltip: getLocalizedOrDefaultLabel(
       'notifications',
       `status.${status}-tooltip${isMultiRecipient ? '-multirecipient' : ''}`,
-      defaultTooltip,
+      undefined,
       filteredData
     ),
     description: getLocalizedOrDefaultLabel(
       'notifications',
       `status.${status}-description${isMultiRecipient ? '-multirecipient' : ''}`,
-      defaultDescription,
+      undefined,
       filteredData
     ),
   };
+}
+
+/*
+ * Returns the mapping between current notification delivered status, label and descriptive message for PA
+ * @param  {NotificationStatus} status
+ * @returns object
+ */
+function getNotificationDeliveredInfosForPA(
+  statusInfos: StatusInfo,
+  statusObject: NotificationStatusHistory | undefined,
+  statusHistory: Array<NotificationStatusHistory>
+): StatusInfo {
+  if (!statusObject) {
+    return statusInfos;
+  }
+  // the timeline event that tells us if there is a stock and its status can be in the DELIVERED status and also in the DELIVERING one.
+  // we have to get the DELIVERING STATUS and get only those events that are cronologically before the DELIVERED status
+  // for each status we have the activeFrom property that is the date from witch the status starts to take effect
+  const activeFrom = new Date(statusObject.activeFrom);
+  // the timeline can be corrected and so we can have multiple DELIVERED statuses
+  // in this case we have to get only those DELIVERING events that are releated to the current DELIVERED status
+  // to do this, we have to check if there is a DELIVERED status next to the current (beacuse the events are ordered from most recent to the oldest)
+  const nextDeliveredStatus = statusHistory.find(
+    (status) =>
+      status.status === NotificationStatus.DELIVERED && new Date(status.activeFrom) < activeFrom
+  );
+  // calc when the next status starts to take effect
+  const activeTo = nextDeliveredStatus ? new Date(nextDeliveredStatus.activeFrom) : null;
+  // get delivering status and filter out those events that are from the date of delivered status activation
+  // and the date of the next status activation (i.e. the date from witch the delivered status ends to take effect)
+  const deliveringStatus = statusHistory.find(
+    (history) => history.status === NotificationStatus.DELIVERING
+  );
+  const deliveringStepsBeforeDelivered = deliveringStatus?.steps?.filter((step) =>
+    activeTo
+      ? new Date(step.timestamp) <= activeFrom && new Date(step.timestamp) > activeTo
+      : new Date(step.timestamp) <= activeFrom
+  );
+  const stepsToCheck = [...(statusObject.steps || []), ...(deliveringStepsBeforeDelivered ?? [])];
+  const isStock = stepsToCheck.find(
+    (step) =>
+      (step.category === TimelineCategory.SEND_ANALOG_PROGRESS ||
+        step.category === TimelineCategory.SEND_ANALOG_FEEDBACK) &&
+      AnalogDeliveryCodeStock.has((step.details as SendPaperDetails).deliveryDetailCode ?? '')
+  );
+  const isWithdrawnStock = stepsToCheck.find(
+    (step) =>
+      (step.category === TimelineCategory.SEND_ANALOG_PROGRESS ||
+        step.category === TimelineCategory.SEND_ANALOG_FEEDBACK) &&
+      AnalogDeliveryCodeWithdrawnStock.has(
+        (step.details as SendPaperDetails).deliveryDetailCode ?? ''
+      )
+  );
+  const isExpiredStock = stepsToCheck.find(
+    (step) =>
+      (step.category === TimelineCategory.SEND_ANALOG_PROGRESS ||
+        step.category === TimelineCategory.SEND_ANALOG_FEEDBACK) &&
+      AnalogDeliveryCodeExpiredStock.has(
+        (step.details as SendPaperDetails).deliveryDetailCode ?? ''
+      )
+  );
+  // case giacenza
+  if (isExpiredStock) {
+    // case expired stock
+    statusInfos.label = getLocalizedOrDefaultLabel(
+      'notifications',
+      'status.delivered-expired-stock'
+    );
+    statusInfos.description = getLocalizedOrDefaultLabel(
+      'notifications',
+      'status.delivered-expired-stock-description'
+    );
+  } else if (isWithdrawnStock) {
+    // case withdraw stock
+    statusInfos.label = getLocalizedOrDefaultLabel(
+      'notifications',
+      'status.delivered-withdrawn-stock'
+    );
+    statusInfos.description = getLocalizedOrDefaultLabel(
+      'notifications',
+      'status.delivered-withdrawn-stock-description'
+    );
+  } else if (isStock) {
+    statusInfos.label = getLocalizedOrDefaultLabel('notifications', 'status.delivered-stock');
+    statusInfos.description = getLocalizedOrDefaultLabel(
+      'notifications',
+      'status.delivered-stock-description'
+    );
+  } else {
+    statusInfos.label = getLocalizedOrDefaultLabel(
+      'notifications',
+      'status.delivered-monorecipient'
+    );
+  }
+  return statusInfos;
 }
 
 /**
@@ -85,7 +178,11 @@ function localizeStatus(
  */
 export function getNotificationStatusInfos(
   status: NotificationStatus | NotificationStatusHistory,
-  options?: { recipients: Array<NotificationDetailRecipient | string> }
+  options?: {
+    statusHistory?: Array<NotificationStatusHistory>;
+    recipients: Array<NotificationDetailRecipient | string>;
+    isParty?: boolean;
+  }
 ): {
   color: 'warning' | 'error' | 'success' | 'info' | 'default' | 'primary' | 'secondary' | undefined;
   label: string;
@@ -119,30 +216,26 @@ export function getNotificationStatusInfos(
 
   switch (actualStatus) {
     case NotificationStatus.DELIVERED: {
-      const statusInfos = localizeStatus(
-        'delivered',
-        'Consegnata',
-        `La notifica è stata consegnata`,
-        'La notifica è stata consegnata.',
-        { isMultiRecipient }
-      );
+      const statusInfos = localizeStatus('delivered', { isMultiRecipient });
       // if the deliveryMode is defined, then change the description for a more specific one ...
       const deliveryMode = statusObject?.deliveryMode;
       // ... only for single-recipient notifications!
-      if (deliveryMode && !isMultiRecipient) {
+      if (deliveryMode && !isMultiRecipient && !options?.isParty) {
         const deliveryModeDescription = getLocalizedOrDefaultLabel(
           'notifications',
-          `status.deliveryMode.${deliveryMode}`,
-          `${deliveryMode}`
+          `status.deliveryMode.${deliveryMode}`
         );
         statusInfos.description = getLocalizedOrDefaultLabel(
           'notifications',
           'status.delivered-description-with-delivery-mode',
-          `La notifica è stata consegnata per via ${
-            deliveryMode === NotificationDeliveryMode.ANALOG ? 'analogica' : 'digitale'
-          }.`,
+          undefined,
           { deliveryMode: deliveryModeDescription }
         );
+      }
+      // if it is a PA, change the title and the description
+      // ... only for single-recipient notifications!
+      if (options?.isParty && !isMultiRecipient) {
+        getNotificationDeliveredInfosForPA(statusInfos, statusObject, options.statusHistory || []);
       }
       // set the color at the end to avoid a type error since the color is defined as an union among some well-known strings
       return { color: 'default', ...statusInfos };
@@ -150,54 +243,27 @@ export function getNotificationStatusInfos(
     case NotificationStatus.DELIVERING:
       return {
         color: 'default',
-        ...localizeStatus(
-          'delivering',
-          'Invio in corso',
-          "L'invio della notifica è in corso",
-          "L'invio della notifica è in corso"
-        ),
+        ...localizeStatus('delivering'),
       };
     case NotificationStatus.UNREACHABLE:
       return {
         color: 'error',
-        ...localizeStatus(
-          'unreachable',
-          'Destinatario irreperibile',
-          'Il destinatario non è reperibile',
-          'Il destinatario non è reperibile',
-          { isMultiRecipient }
-        ),
+        ...localizeStatus('unreachable', { isMultiRecipient }),
       };
     case NotificationStatus.PAID:
       return {
         color: 'success',
-        ...localizeStatus(
-          'paid',
-          'Pagata',
-          'Il destinatario ha pagato i costi della notifica',
-          'Il destinatario ha pagato i costi della notifica'
-        ),
+        ...localizeStatus('paid'),
       };
     case NotificationStatus.ACCEPTED:
       return {
         color: 'default',
-        ...localizeStatus(
-          'accepted',
-          'Depositata',
-          "L'ente ha depositato la notifica",
-          "L'ente ha depositato la notifica"
-        ),
+        ...localizeStatus('accepted'),
       };
     case NotificationStatus.EFFECTIVE_DATE:
       return {
         color: 'info',
-        ...localizeStatus(
-          'effective-date',
-          'Perfezionata per decorrenza termini',
-          'Il destinatario non ha letto la notifica',
-          'Il destinatario non ha letto la notifica entro il termine stabilito',
-          { isMultiRecipient }
-        ),
+        ...localizeStatus('effective-date', { isMultiRecipient }),
       };
     case NotificationStatus.VIEWED:
       if (statusObject?.recipient) {
@@ -210,54 +276,27 @@ export function getNotificationStatusInfos(
       }
       return {
         color: 'success',
-        ...localizeStatus(
-          'viewed',
-          'Avvenuto accesso',
-          `Il ${subject} ha letto la notifica`,
-          `Il ${subject} ha letto la notifica`,
-          { subject, isMultiRecipient }
-        ),
+        ...localizeStatus('viewed', { subject, isMultiRecipient }),
       };
     case NotificationStatus.CANCELLED:
       return {
         color: 'warning',
-        ...localizeStatus(
-          'canceled',
-          'Annullata',
-          "L'ente ha annullato l'invio della notifica",
-          "L'ente ha annullato l'invio della notifica"
-        ),
+        ...localizeStatus('canceled'),
       };
     case NotificationStatus.CANCELLATION_IN_PROGRESS:
       return {
         color: 'warning',
-        ...localizeStatus(
-          'cancellation-in-progress',
-          'Annullata',
-          'Annullamento in corso. Lo stato sarà aggiornato a breve.',
-          'Annullamento in corso. Lo stato sarà aggiornato a breve.'
-        ),
+        ...localizeStatus('cancellation-in-progress'),
       };
     case NotificationStatus.RETURNED_TO_SENDER:
       return {
         color: 'warning',
-        ...localizeStatus(
-          'returned-to-sender',
-          'Resa al mittente',
-          `Il destinatario risulta deceduto.`,
-          `Il destinatario risulta deceduto.`,
-          { isMultiRecipient }
-        ),
+        ...localizeStatus('returned-to-sender', { isMultiRecipient }),
       };
     case NotificationStatus.NOTIFICATION_TIMELINE_REWORKED:
       return {
         color: 'warning',
-        ...localizeStatus(
-          'notification-timeline-reworked',
-          'Evento o più eventi aggiornati',
-          `Evento o più eventi aggiornati`,
-          `Evento o più eventi aggiornati`
-        ),
+        ...localizeStatus('notification-timeline-reworked'),
       };
     default:
       return {
@@ -272,51 +311,39 @@ export function getNotificationStatusInfos(
 export const getNotificationAllowedStatus = () => [
   {
     value: 'All',
-    label: getLocalizedOrDefaultLabel('notifications', 'status.all', 'Tutti gli stati'),
+    label: getLocalizedOrDefaultLabel('notifications', 'status.all'),
   },
   {
     value: NotificationStatus.ACCEPTED,
-    label: getLocalizedOrDefaultLabel('notifications', 'status.accepted', 'Depositata'),
+    label: getLocalizedOrDefaultLabel('notifications', 'status.accepted'),
   },
   {
     value: NotificationStatus.DELIVERING,
-    label: getLocalizedOrDefaultLabel('notifications', 'status.delivering', 'Invio in corso'),
+    label: getLocalizedOrDefaultLabel('notifications', 'status.delivering'),
   },
   {
     value: NotificationStatus.DELIVERED,
-    label: getLocalizedOrDefaultLabel('notifications', 'status.delivered', 'Consegnata'),
+    label: getLocalizedOrDefaultLabel('notifications', 'status.delivered'),
   },
   {
     value: NotificationStatus.EFFECTIVE_DATE,
-    label: getLocalizedOrDefaultLabel(
-      'notifications',
-      'status.effective-date',
-      'Perfezionata per decorrenza termini'
-    ),
+    label: getLocalizedOrDefaultLabel('notifications', 'status.effective-date'),
   },
   {
     value: NotificationStatus.VIEWED,
-    label: getLocalizedOrDefaultLabel('notifications', 'status.viewed', 'Avvenuto accesso'),
+    label: getLocalizedOrDefaultLabel('notifications', 'status.viewed'),
   },
   {
     value: NotificationStatus.CANCELLED,
-    label: getLocalizedOrDefaultLabel('notifications', 'status.canceled', 'Annullata'),
+    label: getLocalizedOrDefaultLabel('notifications', 'status.canceled'),
   },
   {
     value: NotificationStatus.UNREACHABLE,
-    label: getLocalizedOrDefaultLabel(
-      'notifications',
-      'status.unreachable',
-      'Destinatario irreperibile'
-    ),
+    label: getLocalizedOrDefaultLabel('notifications', 'status.unreachable'),
   },
   {
     value: NotificationStatus.RETURNED_TO_SENDER,
-    label: getLocalizedOrDefaultLabel(
-      'notifications',
-      'status.returned-to-sender',
-      'Resa al mittente'
-    ),
+    label: getLocalizedOrDefaultLabel('notifications', 'status.returned-to-sender'),
   },
 ];