+++ packages/pn-pa-webapp/src/redux/dashboard/__test__/reducers.test.ts
@@ -42,8 +42,8 @@ describe('Dashboard redux state tests', () => {
       loading: false,
       notifications: [],
       filters: {
-        startDate: tenYearsAgo,
-        endDate: today,
+        startDate: undefined,
+        endDate: undefined,
         status: '',
         recipientId: '',
         iunMatch: '',